+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [1.1.0] - 2025-09-22
+
+### New Features
+
+- DNS zone and records configuration
+- Virtual Machines A and CNAME records creation (root)
+
+### Improvements
+
+- [BREAKING] Refactored resource group resource configuration (root)
+
 ## [1.0.0] - 2025-09-15
 
 ### New Features